fix: addressed PR review comments
