+++ app/dapp/support/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./CommunityCard";
+export * from "./GuidesCard";
